+++ example/lib/attached_toast/attached_toast.dart
@@ -10,34 +10,56 @@ class AttachedToast extends StatefulWidget {
 class _AttachedToastState extends State<AttachedToast> {
   BuildContext targetContext;
   Offset target;
-  double verticalOffset = 24;
+  double verticalOffset = 0;
+  double horizontalOffset = 0;
   int second = 4;
-  PreferDirection preferDirection = PreferDirection.Below;
+  PreferDirection preferDirection = PreferDirection.topCenter;
   bool ignoreContentClick = false;
   bool onlyOne = true;
   bool allowClick = true;
+  bool enableSafeArea = true;
   int backgroundColor = 0x00000000;
 
-  double buttonAlign=0;
+  double buttonAlign = 0;
 
   CancelFunc show({BuildContext context, Offset target}) {
     return BotToast.showAttachedWidget(
         target: target,
         targetContext: context,
         verticalOffset: verticalOffset,
+        horizontalOffset: horizontalOffset,
         duration: Duration(seconds: second),
         preferDirection: preferDirection,
         ignoreContentClick: ignoreContentClick,
         onlyOne: onlyOne,
         allowClick: allowClick,
+        enableSafeArea: enableSafeArea,
         backgroundColor: Color(backgroundColor),
         attachedBuilder: (cancel) => (Card(
-              child: IconButton(
-                padding: EdgeInsets.all(20),
-                onPressed: (){
-                  BotToast.showSimpleNotification(title: "Let's go travel together.😘");
-                },
-                icon: Icon(Icons.favorite,color: Colors.redAccent,),
+          color: Colors.amber,
+              child: Container(
+                padding: EdgeInsets.symmetric(horizontal: 8),
+//                width: 200,
+                child: Row(
+                  mainAxisSize: MainAxisSize.min,
+                  children: <Widget>[
+                    FlatButton.icon(
+                      padding: EdgeInsets.all(5),
+                      onPressed: (){
+                        BotToast.showSimpleNotification(title: "Let's go travel together.😘");
+                      },
+                      label: Text("favorite"),
+                      icon: Icon(Icons.favorite,color: Colors.redAccent),
+                    ),
+                    FlatButton.icon(
+                      padding: EdgeInsets.all(5),
+                      onPressed: (){
+                        BotToast.showSimpleNotification(title: "Thank you for liking me.😝");
+                      },
+                      label: Text("bookmark"),
+                      icon: Icon(Icons.bookmark,color: Colors.redAccent),
+                    )                  ],
+                ),
               ),
             )));
   }
@@ -56,20 +78,18 @@ class _AttachedToastState extends State<AttachedToast> {
             children: <Widget>[
               GestureDetector(
                 onTapDown: (details) {
-                  setState(() {
-                    verticalOffset=0;
-
-                  });
                   show(target: details.globalPosition);
                 },
-                child: Container(
-                  color: Colors.redAccent,
-                  height: 100,
-                  width: 300,
-                  alignment: Alignment.center,
-                  child: Text(
-                    "click me",
-                    style: TextStyle(color: Colors.white, fontSize: 20),
+                child: Card(
+                  margin: EdgeInsets.symmetric(horizontal: 25),
+                  color: Colors.black45,
+                  child: Container(
+                    height: 200,
+                    alignment: Alignment.center,
+                    child: Text(
+                      "click me",
+                      style: TextStyle(color: Colors.white, fontSize: 20),
+                    ),
                   ),
                 ),
               ),
@@ -100,18 +120,243 @@ class _AttachedToastState extends State<AttachedToast> {
                   },
                 ),
               ),
-              SwitchListTile(
-                value: preferDirection == PreferDirection.Below,
-                onChanged: (value) {
-                  setState(() {
-                    preferDirection =
-                        value ? PreferDirection.Below : PreferDirection.Upside;
-                  });
-                },
-                title: Text("preferDirection: $preferDirection"),
+              Text("$preferDirection",style: TextStyle(fontSize: 15),),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                    child: SizedBox.shrink(),
+                  ),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.topLeft,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.topCenter,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.topRight,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Spacer(),
+                ],
               ),
-
-
+              Row(
+                children: <Widget>[
+                  Spacer(),
+                  Expanded(
+                      child: Text(
+                    "topLeft",
+                    textAlign: TextAlign.center,
+                  )),
+                  Expanded(
+                      child: Text(
+                    "topCenter",
+                    textAlign: TextAlign.center,
+                  )),
+                  Expanded(
+                      child: Text(
+                    "topRight",
+                    textAlign: TextAlign.center,
+                  )),
+                  Spacer(),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.leftTop,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.rightTop,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Text(
+                    "leftTop",
+                    textAlign: TextAlign.center,
+                  )),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Text(
+                    "rightTop",
+                    textAlign: TextAlign.center,
+                  )),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.leftCenter,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.rightCenter,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Text(
+                    "leftCenter",
+                    textAlign: TextAlign.center,
+                  )),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Text(
+                    "rightCenter",
+                    textAlign: TextAlign.center,
+                  )),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.leftBottom,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.rightBottom,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Expanded(
+                      child: Text(
+                    "leftBottom",
+                    textAlign: TextAlign.center,
+                  )),
+                  Spacer(),
+                  Spacer(),
+                  Spacer(),
+                  Expanded(
+                      child: Text(
+                    "rightBottom",
+                    textAlign: TextAlign.center,
+                  )),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Spacer(),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.bottomLeft,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.bottomCenter,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Expanded(
+                      child: Radio(
+                          value: PreferDirection.bottomRight,
+                          groupValue: preferDirection,
+                          onChanged: (value) {
+                            setState(() {
+                              preferDirection = value;
+                            });
+                          })),
+                  Spacer(),
+                ],
+              ),
+              Row(
+                children: <Widget>[
+                  Spacer(),
+                  Expanded(
+                      child: Text(
+                    "bottomLeft",
+                    textAlign: TextAlign.center,
+                  )),
+                  Expanded(
+                      child: Text(
+                    "bottomCenter",
+                    textAlign: TextAlign.center,
+                  )),
+                  Expanded(
+                      child: Text(
+                    "bottomRight",
+                    textAlign: TextAlign.center,
+                  )),
+                  Spacer(),
+                ],
+              ),
               SwitchListTile(
                 value: ignoreContentClick,
                 onChanged: (value) {
@@ -139,6 +384,15 @@ class _AttachedToastState extends State<AttachedToast> {
                 },
                 title: Text("allowClick: "),
               ),
+              SwitchListTile(
+                value: enableSafeArea,
+                onChanged: (value) {
+                  setState(() {
+                    enableSafeArea = value;
+                  });
+                },
+                title: Text("enableSafeArea: "),
+              ),
               ListTile(
                 title: Text("duration:   ${second}s"),
                 trailing: CupertinoSlider(
@@ -161,6 +415,20 @@ class _AttachedToastState extends State<AttachedToast> {
                   onChanged: (double value) {
                     setState(() {
                       verticalOffset = value;
+                    });
+                  },
+                ),
+              ),
+              ListTile(
+                title:
+                    Text("horizontalOffset:   ${horizontalOffset.toString()}"),
+                trailing: CupertinoSlider(
+                  min: 0,
+                  max: 40,
+                  value: horizontalOffset.toDouble(),
+                  onChanged: (double value) {
+                    setState(() {
+                      horizontalOffset = value;
                     });
                   },
                 ),